+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+- 2020.12.28, 00:25 -- PD-3504: Popup based on `ModalPortal` (`updateGeometry` and other updates `windowDomNode` passed from `ModalPortal`). (With extra debug logging.)
+- 2020.12.28, 23:00 -- PD-3504: ModalPortal: exported `passModalPortalProps` with list of public props, set `modalType` property with `props.type` (from Modal or Popup). (With debug logging.)
 - 2020.12.26, 23:09 -- PD-3504: ModalPortal: inplemented generic modal wrapper code, layout & styles form original `Modal`, Modal: Renamed Modal to ModalWindow.
 - 2020.12.25, 17:30 -- ModalPortal: Выделение общего (для Modal и Popup) компонента ModalPortal, отвечающего за функционирование интерактивных элементов вне общего dom-дерева приложения.
 - 2020.12.25, 00:56 -- Modal & Loader themes updated.
@@ -55,5 +57,5 @@
 - 2020.10.05, 20:38 -- v.0.0.15: Extracted `demoSupport.jsx` routines, fixture selection from url query (parameter `fixture`), added avaialble fixtures list if fixture to display is not specified, added DemoWrapper & demoTitle exportable ficture parameters for demo rendering.
 
 <!--
- @changed 2020.12.26, 23:13
+ @changed 2020.12.28, 00:26
 -->